created basic game setup
A GameManger has been created that manages game states and holds the
major components.
A LevelManager has been created that loads levels.
A Level class has been created that loads and holds GameObjects of
intereset for the interaction logic.
A LevelConfig class holds information about a level such as the maximal
number of player units allowed.
An ImpManager has been created that manages the imps in a level.
An ImpController has been created that manages the behavior of an imp.